+++ src/main/java/com/boot/ict05_final_admin/domain/inventory/dto/MaterialWriteFormDTO.java
@@ -38,7 +38,7 @@ public class MaterialWriteFormDTO {
 
     /** 판매단위 → 기본단위 변환비율 (필수) */
     @NotNull(message = "변환비율을 입력해주세요")
-    private Double conversionRate;
+    private Integer conversionRate;
 
     /** 공급업체명 */
     private String supplier;
